requirements.txt

langchain==1.0.7
langchain-community==0.4.1
langchain-google-genai==3.0.3
langchain-chroma==0.1.4
chromadb==0.5.5

supabase-py==2.4.0
python-dotenv==1.0.1

feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

fastapi==0.115.0
uvicorn[standard]==0.30.6

streamlit==1.38.0
